Feat/4 ir module

See merge request softbdltd/nise3/nise3-organization-management-api-service!336

+++ app/Http/Controllers/FourIRGuidelineController.php
@@ -42,7 +42,7 @@ class FourIRGuidelineController extends Controller
      */
     public function getList(Request $request): JsonResponse
     {
-        //$this->authorize('viewAny', FourIRGuideline::class);
+        $this->authorize('viewAny', FourIRGuideline::class);
 
         $filter = $this->fourIRGuidelineService->filterValidator($request)->validate();
         $response = $this->fourIRGuidelineService->getFourIRGuidelineList($filter, $this->startTime);
@@ -57,7 +57,7 @@ class FourIRGuidelineController extends Controller
     public function read(int $id): JsonResponse
     {
         $guideline = $this->fourIRGuidelineService->getOneGuideline($id);
-        // $this->authorize('view', $rank);
+        $this->authorize('view', $guideline);
         $response = [
             "data" => $guideline,
             "_response_status" => [
@@ -80,7 +80,7 @@ class FourIRGuidelineController extends Controller
      */
     function store(Request $request): JsonResponse
     {
-        //$this->authorize('create', FourIRGuideline::class);
+        $this->authorize('create', FourIRGuideline::class);
         $validated = $this->fourIRGuidelineService->validator($request)->validate();
         $data = $this->fourIRGuidelineService->store($validated);
 
@@ -108,7 +108,7 @@ class FourIRGuidelineController extends Controller
     public function update(Request $request, int $id): JsonResponse
     {
         $fourIrGuideline = FourIRGuideline::findOrFail($id);
-        // $this->authorize('update', $fourIrGuideline);
+         $this->authorize('update', $fourIrGuideline);
         $validated = $this->fourIRGuidelineService->validator($request, $id)->validate();
         $data = $this->fourIRGuidelineService->update($fourIrGuideline, $validated);
 
@@ -135,7 +135,7 @@ class FourIRGuidelineController extends Controller
     public function destroy(int $id): JsonResponse
     {
         $fourIrGuideline = FourIRGuideline::findOrFail($id);
-//        $this->authorize('delete', $fourIrGuideline);
+        $this->authorize('delete', $fourIrGuideline);
         $this->fourIRGuidelineService->destroy($fourIrGuideline);
         $response = [
             '_response_status' => [
